meeting done

+++ Apartment_Admin/src/main/java/com/apartment/models/EmployeePosition.java
@@ -1,12 +1,8 @@
 package com.apartment.models;
 
-import java.security.Timestamp;
-import java.util.Date;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
-
 import javax.persistence.Id;
 
 import lombok.Getter;
@@ -19,7 +15,7 @@ public class EmployeePosition {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int employeePositionId;
+	private Long employeePositionId;
 	private String positionName;
 	private double salary;
 	private String responsibilities;
