+++ Apartment_Admin/src/main/java/com/apartment/repo/EmployeePositionRepo.java
@@ -3,10 +3,9 @@ package com.apartment.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.apartment.models.Election;
 import com.apartment.models.EmployeePosition;
 
 @Repository
-public interface EmployeePositionRepo extends JpaRepository <EmployeePosition,Integer> {
+public interface EmployeePositionRepo extends JpaRepository <EmployeePosition,Long> {
 
 }
